This is an advance cat&dog classifier using subset of kaggle dataset
It uses only 4000 images out of 50,000 images provided in the kaggle tournament of CatvsDog held in 2003 and still manages to predict with 96% accuracy, while the team in the kaggle who won the first place managed to predict 97% accuracy using 50,000 images. The main reason behind this optimum prediction using such small subset is data augmentation and transfer learning.